requirements.txt
asgiref==3.11.1                                                                             
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==6.0.1
certifi==2026.4.22
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
constantly==23.10.4
cryptography==47.0.0
daphne==4.2.1
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
msgpack==1.1.2
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.1.0
python-dotenv==1.2.2
razorpay==2.0.1
redis==7.4.0
requests==2.33.1
service-identity==24.2.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.2
ujson==5.12.0
urllib3==2.6.3
zope.interface==8.4